+++ roadgraph_builder/routing/__init__.py
@@ -6,6 +6,7 @@ from roadgraph_builder.routing.geojson_export import (
     write_reachability_geojson,
     write_route_geojson,
 )
+from roadgraph_builder.routing.edge_index import EdgeProjection, EdgeProjectionIndex
 from roadgraph_builder.routing.hmm_match import HmmMatch, hmm_match_trajectory
 from roadgraph_builder.routing.map_match import SnappedPoint, coverage_stats, snap_trajectory_to_graph
 from roadgraph_builder.routing.nearest import NearestNode, nearest_node
@@ -26,6 +27,8 @@ from roadgraph_builder.routing.trip_reconstruction import Trip, reconstruct_trip
 
 __all__ = [
     "HmmMatch",
+    "EdgeProjection",
+    "EdgeProjectionIndex",
     "NearestNode",
     "ReachabilityAnalyzer",
     "ReachabilityResult",